requirements.txt
langchain_core==1.0.7
langchain-anthropic==1.0.0
langchain-google-genai==3.0.0
langchain-openai==1.0.1
python-dotenv==1.0.1
tree_sitter==0.20.1
tree_sitter_languages==1.10.2
